src/requirements/dev.txt
django-oscar==2.0.4
psycopg2==2.8.4
django-widget-tweaks==1.4.5
django-tables2==1.21.2
Django==2.2.11
sorl-thumbnail==12.6.3
django-simple-captcha
pycountry==19.8.18
celery==4.4.1